[step4] refactor MVC 패키지 변경
