+++ src/main/java/model/BonusBall.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 public class BonusBall {
