+++ src/main/java/model/LottoNo.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import java.util.HashMap;
 import java.util.Map;
