+++ src/main/java/model/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import java.util.Collections;
 import java.util.HashSet;
