+++ src/main/java/model/Matches.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import java.util.Arrays;
