+++ src/main/java/model/MatchesStatus.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import java.math.BigDecimal;
