+++ src/test/java/model/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
